# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Go语言和fiber框架的高性能高并发的Web项目骨架",
        "title": "Go-skeleton API Docs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9528",
    "basePath": "/",
    "paths": {
...
        "/farm/account/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "创建农场账号",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/account/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "删除农场账号",
                "parameters": [
                    {
                        "description": "删除参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/account/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "农场账号详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/account/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "农场账号列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称/编码/QQ关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态 1正常 2禁用",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运行状态",
                        "name": "runStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/account/modify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "更新农场账号",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/account/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "启动农场账号",
                "parameters": [
                    {
                        "description": "账号ID",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountIDReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/account/stop": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场账号"
                ],
                "summary": "停止农场账号",
                "parameters": [
                    {
                        "description": "账号ID",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountIDReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/automation/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场自动化"
                ],
                "summary": "自动化配置详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID",
                        "name": "accountId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/automation/modify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场自动化"
                ],
                "summary": "修改自动化配置",
                "parameters": [
                    {
                        "description": "配置参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/automation.AutomationModifyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "account.AccountCreateReq": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "maxLength": 512
                },
                "code": {
                    "type": "string",
                    "maxLength": 2048
                },
                "name": {
                    "type": "string",
                    "maxLength": 64
                },
                "platform": {
                    "type": "string",
                    "enum": [
                        "qq",
                        "wx"
                    ]
                },
                "qq": {
                    "type": "string",
                    "maxLength": 32
                },
                "remark": {
                    "type": "string",
                    "maxLength": 255
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "uin": {
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
        "account.AccountDeleteReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "account.AccountIDReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "account.AccountUpdateReq": {
            "type": "object",
            "required": [
                "code",
                "id",
                "status"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "maxLength": 512
                },
                "code": {
                    "type": "string",
                    "maxLength": 2048
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 64
                },
                "platform": {
                    "type": "string",
                    "enum": [
                        "qq",
                        "wx"
                    ]
                },
                "qq": {
                    "type": "string",
                    "maxLength": 32
                },
                "remark": {
                    "type": "string",
                    "maxLength": 255
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "uin": {
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
        "automation.AutomationModifyReq": {
            "type": "object",
            "required": [
                "accountId"
            ],
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "automation": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "bagSeedFallbackStrategy": {
                    "type": "string"
                },
                "bagSeedPriority": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "configJson": {
                    "description": "兼容旧客户端",
                    "type": "string"
                },
                "fertilizerBuyCheckIntervalMinutes": {
                    "type": "integer"
                },
                "fertilizerBuyNormalCount": {
                    "type": "integer"
                },
                "fertilizerBuyNormalThresholdHours": {
                    "type": "integer"
                },
                "fertilizerBuyOrganicCount": {
                    "type": "integer"
                },
                "fertilizerBuyOrganicThresholdHours": {
                    "type": "integer"
                },
                "friendBlacklist": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "friendQuietHours": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "intervals": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "plantBlacklist": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "plantDelaySeconds": {
                    "type": "integer"
                },
                "plantOrderRandom": {
                    "type": "boolean"
                },
                "plantingStrategy": {
                    "type": "string"
                },
                "preferredSeedId": {
                    "type": "integer"
                },
                "stealDelaySeconds": {
                    "type": "integer"
                }
            }
        },
...
        "response.Code": {
            "type": "integer",
            "enum": [
                0,
                10001,
                10002,
                10003,
                10004,
                10005,
                10006,
                10007,
                10008
            ],
            "x-enum-varnames": [
                "Success",
                "Failed",
                "UnAuthed",
                "AuthExpired",
                "InternalErr",
                "RequestMethodErr",
                "RequestParamErr",
                "RequestFormErr",
                "AccountInfoErr"
            ]
        },
        "response.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.Code"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}